Add tracer.js
This is has nothing directly related to lsn or ana but it is a possibly
useful tool to determine how a function interacts with its arguments. In
particular, this will be useful for stream functions to be run so that
their results can be analyzed.

Eventually, the plan is for ana programs to be run so that an engine
attach source and sink streams for various side effects.