Add helper pathRef fns and shared pathRefs
Add:
- pathRefToString
- pathRefToArray
- arrayToPathRef
- concatPathRefs

```javascript
pathRefToString($.a.b) // '$.a.b'
pathRefToArray($.a.b) // ['a', 'b']
arrayToPathRef([a, b]) // $.a.b
concatPathRefs($.a, $.b) // $.a.b
```

pathRefs are reused when possible so that no new pathRef is allocated
for the same path. This property is sharef for their arrays.

```javascript
x = $.a.b
y = $.a
x === y.b // true
pathRefToArray(x) === pathRefToArray(y.b)
```